Inline GameView observer helpers
